refactor: UrlRow, Details
